workers/datasets_based/pyproject.toml
[tool.poetry]
authors = ["Sylvain Lesage <sylvain.lesage@huggingface.co>"]
description = "Worker for processing steps that need the datasets library"
name = "datasets_based"
version = "1.0.0"
license = "Apache-2.0"

[tool.poetry.dependencies]
Pillow = "^9.3.0"
PyICU = "^2.7.4"
aiohttp = "^3.7.4.post0"
apache-beam = "2.41.0" # ^2 gives a InvalidWheelName error because it tries to install 2.42 that has not been released...
bs4 = "^0.0.1"
conllu = "^4.4.1"
datasets = { extras = ["audio", "vision"], version = "~2.7.1" }
gdown = "^4.2.0"
huggingface-hub = "^0.11.0"
kenlm = { url = "https://github.com/kpu/kenlm/archive/master.zip" }
kss = "^2.6.0"
libcommon = { path = "../../libs/libcommon/dist/libcommon-0.5.8-py3-none-any.whl", develop = false }
lm-dataformat = "^0.0.20"
lxml = "^4.9.1"
nlp = "^0.4.0"
nltk = "^3.6.5"
openpyxl = "^3.0.9"
py7zr = "^0.20.1"
pydub = "^0.25.1"
python = "3.9.15"
rarfile = "^4.0"
scikit-learn = "^1.0"
tensorflow = {version = "^2.9.1", markers = "sys_platform != 'darwin' or platform_machine != 'arm64'"}
tensorflow-macos = {version = "^2.9.1", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
tfrecord = "^1.14.1"
torchaudio = "^0.10.1"
transformers = "^4.11.3"
trec-car-tools = { path = "vendors/trec-car-tools/python3" }
typer = "^0.4.0"
wget = "^3.2"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.0"
black = "^22.1.0"
flake8 = "^3.9.2"
isort = "^5.9.3"
mypy = "0.812"
pandas-stubs = "^1.5.2.221124"
poetryup = "^0.3.8"
pytest = "^7.2.0"
pytest-cov = "^2.12.1"
pip-audit = "^2.4.6"
types-requests = "^2.28.11"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
# addopts = "-k 'wip'"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "wip: tests being developed"
]

[tool.coverage.run]
source = ["datasets_based"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 119
preview = true

[tool.mypy]
exclude = 'vendors'
strict = true

[[tool.mypy.overrides]]
# could be solved in datasets by adding __add__ to the __init__.py file
implicit_reexport = true
module = "datasets"
